+++ src/utils/const.ts
@@ -1,1 +1,0 @@
-export const MODULE_ORDER = 999;